requirements.txt
appdirs==1.4.3
appnope==0.1.0
beautifulsoup4==4.5.3
decorator==4.0.11
Django==1.10.6
django-countries==4.5
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
django-environ==0.4.3
django-extensions==1.7.8
django-isbn-field==0.3
django-livereload-server==0.2.3
django-redis==4.8.0
djangorestframework==3.6.2
flake8==3.3.0
flake8-blind-except==0.1.1
flake8-debugger==1.4.0
ipdb==0.10.3
ipython==5.3.0
ipython-genutils==0.2.0
isort==4.2.5
jedi==0.10.2
mccabe==0.6.1
olefile==0.44
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.1
pycodestyle==2.3.1
pycountry==17.1.8
pydot==1.2.3
pyflakes==1.5.0
Pygments==2.2.0
pyparsing==2.2.0
pyprof2calltree==1.4.0
pysolr==3.6.0
python-stdnum==1.6
redis==2.10.5
requests==2.13.0
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.3
tornado==4.4.3
traitlets==4.3.2
transifex-client==0.12.4
urllib3==1.21.1
uWSGI==2.0.15
wcwidth==0.1.7
Werkzeug==0.12.1
whitenoise==3.3.0
Whoosh==2.7.4
